+++ CHANGELOG.md
@@ -47,6 +47,7 @@ All notable changes to this project will be documented here. The format is based
 - Clicking a line of action now selects the associated muscle (#571)
 - Fixed a bug in the DAE exporter that caused it to throw an 'argument not found' exception (#581)
 - Lines of action rendering now has separate checkboxes for muscle insertion/origin (#577)
+- Stations can now be dragged around (translated) in model editor tab's 3D viewport (#85)
 
 ## [0.3.2] - 2023/01/09